pyproject.toml
# pyproject.toml

# Note: Many fields are autogenerated by 'WIPACrepo/wipac-dev-py-setup-action'.
#       See this repo's GitHub Actions workflow(s) (.github/workflows/*.yaml)
#       and https://github.com/WIPACrepo/wipac-dev-py-setup-action for details.

[project]
dependencies = [
    'PyYAML',
    'asyncache',
    'boto3<1.36',
    'cachetools',
    'certifi',
    'cryptography',
    'htcondor<=24.99.99',
    'httpx',
    'jsonschema',
    'ldap3',
    'motor',
    'prometheus-client',
    'psutil',
    'pyOpenSSL',
    'pymongo',
    'python-dateutil',
    'requests',
    'requests-futures',
    'requests-toolbelt',
    'setproctitle',
    'tornado',
    'unidecode',
    'wipac-dev-tools',
    'wipac-rest-tools>=1.11.5',
]
dynamic = ["version"] # do not edit — autogenerated by wipac-dev-py-setup-action
name = "iceprod" # do not edit — autogenerated by wipac-dev-py-setup-action
description = "IceCube dataset management system" # do not edit — autogenerated by wipac-dev-py-setup-action
readme = "README.rst" # do not edit — autogenerated by wipac-dev-py-setup-action
license = "MIT" # do not edit — autogenerated by wipac-dev-py-setup-action
license-files = ["LICENSE"] # do not edit — autogenerated by wipac-dev-py-setup-action
keywords = [
    'WIPAC',
    'batch',
    'workload',
] # do not edit — autogenerated by wipac-dev-py-setup-action
classifiers = [
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
] # do not edit — autogenerated by wipac-dev-py-setup-action
requires-python = ">=3.11, <3.14" # do not edit — autogenerated by wipac-dev-py-setup-action

[project.optional-dependencies]
tests = [
    'beautifulsoup4',
    'coverage',
    'flake8',
    'flexmock',
    'mock',
    'moto[s3]>=5',
    'pytest-asyncio',
    'pytest-cov',
    'pytest-mock',
    'pytest<8.1',
    'requests-mock',
    'respx',
    'types-cachetools',
    'types-requests',
]
docs = [
    'sphinx',
]
mypy = [
    'beautifulsoup4',
    'coverage',
    'flake8',
    'flexmock',
    'mock',
    'moto[s3]>=5',
    'pytest-asyncio',
    'pytest-cov',
    'pytest-mock',
    'pytest<8.1',
    'requests-mock',
    'respx',
    'sphinx',
    'types-cachetools',
    'types-requests',
] # do not edit — autogenerated by wipac-dev-py-setup-action

[[project.authors]]
name = "WIPAC Developers"
email = "developers@icecube.wisc.edu"

[project.urls] # do not edit — autogenerated by wipac-dev-py-setup-action
Homepage = "https://pypi.org/project/iceprod/"
Tracker = "https://github.com/WIPACrepo/iceprod/issues"
Source = "https://github.com/WIPACrepo/iceprod"

[build-system] # do not edit — autogenerated by wipac-dev-py-setup-action
requires = ["setuptools>=78.1", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find] # do not edit — autogenerated by wipac-dev-py-setup-action
include = ["iceprod", "iceprod.*"]

[tool.setuptools.package-data]
"*" = ["py.typed"] # do not edit — autogenerated by wipac-dev-py-setup-action

[tool.setuptools_scm] # do not edit — autogenerated by wipac-dev-py-setup-action
fallback_version = "UNTAGGED" # only used when SCM metadata is completely unavailable (e.g., no '.git/'). NOTE: in shallow clones without git tags, setuptools-scm will still infer a commit-based version (like '0.1.dev1+g79c9b4b'); this value will NOT be used in that case.

[tool.mypy]
exclude = [
    '^docs',
    '^integration_tests',
    '^tests',
    '^resources',
]
exclude_gitignore = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
